+++ src/constant_medium.h
@@ -19,8 +19,8 @@
 
 struct constant_medium {
     traversable_geometry geom;
-    double neg_inv_density;
-    constant_medium(traversable_geometry boundary, double density)
+    float neg_inv_density;
+    constant_medium(traversable_geometry boundary, float density)
         : geom(boundary)
         , neg_inv_density(-1 / density)
     {
