+++ src/vec3.h
@@ -16,24 +16,24 @@
 
 class vec3 {
 public:
-    double e[3];
+    float e[3];
 
     constexpr vec3()
         : e { 0, 0, 0 }
     {
     }
-    constexpr vec3(double e0, double e1, double e2)
+    constexpr vec3(float e0, float e1, float e2)
         : e { e0, e1, e2 }
     {
     }
 
-    constexpr double x() const { return e[0]; }
-    constexpr double y() const { return e[1]; }
-    constexpr double z() const { return e[2]; }
+    constexpr float x() const { return e[0]; }
+    constexpr float y() const { return e[1]; }
+    constexpr float z() const { return e[2]; }
 
     constexpr vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-    constexpr double operator[](int i) const { return e[i]; }
-    constexpr double &operator[](int i) { return e[i]; }
+    constexpr float operator[](int i) const { return e[i]; }
+    constexpr float &operator[](int i) { return e[i]; }
 
     constexpr vec3 &operator+=(vec3 const &v)
     {
@@ -43,7 +43,7 @@ public:
         return *this;
     }
 
-    constexpr vec3 &operator*=(double t)
+    constexpr vec3 &operator*=(float t)
     {
         e[0] *= t;
         e[1] *= t;
@@ -51,11 +51,11 @@ public:
         return *this;
     }
 
-    constexpr vec3 &operator/=(double t) { return *this *= 1 / t; }
+    constexpr vec3 &operator/=(float t) { return *this *= 1 / t; }
 
-    double length() const { return std::sqrt(length_squared()); }
+    float length() const { return std::sqrt(length_squared()); }
 
-    constexpr double length_squared() const
+    constexpr float length_squared() const
     {
         return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
     }
@@ -89,16 +89,16 @@ constexpr vec3 operator*(vec3 u, vec3 v)
     return vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-constexpr vec3 operator*(double t, vec3 v)
+constexpr vec3 operator*(float t, vec3 v)
 {
     return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-constexpr vec3 operator*(vec3 v, double t) { return t * v; }
+constexpr vec3 operator*(vec3 v, float t) { return t * v; }
 
-constexpr vec3 operator/(vec3 v, double t) { return (1 / t) * v; }
+constexpr vec3 operator/(vec3 v, float t) { return (1 / t) * v; }
 
-constexpr double dot(vec3 u, vec3 v)
+constexpr float dot(vec3 u, vec3 v)
 {
     return u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2];
 }
